requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.46
stripe==14.3.0
tomli==2.4.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
zipp==3.23.0
